Trait bee_protocol::packet::Packet

pub(crate) trait Packet {
    const ID: u8;

    fn size_range() -> Range<usize>;

    fn from_bytes(bytes: &[u8]) -> Self;

    fn size(&self) -> usize;

    fn into_bytes(self, bytes: &mut [u8]);
}

A trait describing the behavior of a packet.

This trait is protocol agnostic and only provides serialization and deserialization to and from byte buffers. It should not be used as is but rather be paired with a higher layer - like a type-length-value encoding - and as such does not provide any bounds check on inputs/outputs buffers.

Associated Constants

const ID: u8

The unique identifier of the packet within the protocol.

Loading content...
Required methods

fn size_range() -> Range<usize>

Returns the size range of the packet as it can be compressed.

fn from_bytes(bytes: &[u8]) -> Self

Deserializes a byte buffer into a packet.

Arguments

bytes - The byte buffer to deserialize from.

Panics

Panics if the provided buffer has an invalid size. The size of the buffer should be within the range returned by the size_range method.

fn size(&self) -> usize

Returns the size of the packet.

fn into_bytes(self, bytes: &mut [u8])

Serializes a packet into a byte buffer.

Arguments

bytes - The byte buffer to serialize into.

Panics

Panics if the provided buffer has an invalid size. The size of the buffer should be equal to the one returned by the size method.

Loading content...